Fix bug where KIF tapped in wrong place
The logic for tapping within a specific rectangle inside a view was broken
Also removed the requirement that a view be marked as an accessibility element in order to match our search. Things like table cells have labels and are perfectly tappable, but don't have the -isAccessibilityElement flag set for some reason.