Add step to tap at a given screen point
In certain situations the accessibility labels don't cut it for being able to find and tap a view, so this adds a "tap at a certain screen point" step to handle more interesting cases. We hit a situation with the built in media picker where Apple's code isn't quite accessible, so this becomes necessary.