Better handle an edge case transition
There is a rare race condition where we can try to tap while a view is in transition due to an animation. In this case we can't get a tappable point and fail. This fix makes it so that we simply wait until the view is in a stable state instead.