Fix the step to test for the absence of a view
There was an issue with the test-for-absence-of-view step where it would incorrectly pass if an animation was in flight because interaction events were being ignored at the application level. This was being interpretted as the view not being available and therefore not present, which is incorrect. I reimplemented the step to not build off of the view-is-present step since much of that step doesn't really make sense for this test, and error messages can be better.