Added a variant of stepToTapViewWithAccessibilityLabel that allows matching a block
Added a variant of stepToTapViewWithAccessibilityLabel that allows
matching a block.  The block returns the name of the label to tap.
This allows for situations where you don't know what the accessibility
label will be until the step is executed, for example, when the label
includes the value of a setting that may be modified by previous
scenarios.