When the accessibility frame is not set, use the view frame
On the iOS Simulator 5.0, there are times when the Accessibility Inspector will
determine the accessibility frame correctly but the accessibilityFrame property
on the element will be {0, 0, 0, 0}. This happens in some cases when the UI is
created programmatically.

As a work-around when running on 5.0 and prior, this change attempts to
duplicate the OS's behavior of guessing the accessibility frame. It assumes
{0, 0, WIDTH, HEIGHT} where WIDTH and HEIGHT are equal to view.frame.width and
view.frame.height.