Future iOS SDK compatibility.
Makes app work with SDK's that do not contain UITouch->_gestureView.
Work around defect in current betas where accessibilityValue may return
an NSAttributedString.
Address issues where an application's keyWindow may not be a member of
the windows array.
Increase the timeout between scrolling to a cell and looking it up.

Note that the latest simulator release has issues with accessibility
and the tests will need to be run on device or with an earlier
simulator.