Resolved scrolling issue.
In a standard application, the run loop is managed by GSRunLoopModal,
which switches between two run loop modes, kCFRunLoopDefaultMode and
UITrackingRunLoopMode, depending on whether or not scrolling is taking
place.  In this test model, we were always running in
kCFRunLoopDefaultMode, meaning that events registered to the tracking
mode were never executed.

This is overcome by tracking when a scrolling gesture recognizer gets
accepted and using the tracking mode to spin the run loop until the
scroll view is done decelerating.