Better error message for failing predicate
When an UIAccessibility element matching a compound predicate cannot be found
KIF tries all combinations of the predicate subpredicates in order to
find the closest match. What is found is then returned as the error
message to give as much information as possible.
Ex: Found element matching predicate 'accessibilityTraits & 8 == 8 AND
accessibilityLabel MATCHES "A"' but not 'accessibilityValue MATCHES
"myValue"''