Use iOS default "Build Active Architecture Only"
Defer to iOS default settings for "Build Active Architecture Only"
which allows multiple architectures to be built with the Debug build
configuration.

Many users of KIF are being forced to fork in order to enable building
for i386 architecture. By removing this restriction, there will be a
larger compatibility with existing applications looking to add
functional testing with KIF.