Remove non-visible table view cells so that they don't stick around.
The naive approach to finding accessibility elements in a table view
results in cells being instantiated and added as subviews of the table
view, despite not being visible. When the table view has thousands of
rows with heavyweight cells this can cause severe memory pressure.
This commit uses an explicit approach to finding accessibility elements
in table views, which allows for proper removal of non-visible cells.