Use scroll view delegate to reliably wait for scrolling to finish
Instead of hard coded arbitrary waits, use scroll view proxy delegate to wait for scrolling to finish.
This removes the need to wait for an arbitrary amount of time (which may not be enough for slower machines, or may be too long for faster ones).
On a relatively new machine, this improves performance for lookups in scroll views (table views and collection views) by 2x and more.