Update Test cases for [hopefully] better travis results
After running the test cases locally on both hardware and sim
it was noticed that some tests needed a bit more waiting to run
reliably. Running too fast on robust platform.

Notably:
Tests that type into input fields:
* In "(void)expectView:(UIView *)view toContainText:(NSString *)expectedResult"
the timeout was set to 1.0. This bumps it up to 2.0. Seems to stabilize field tests.
* Predictive text was getting in the way of typing the string, "Test-Driven Devlopment".
The text was getting auto-corrected.  This PR changes it to "Testing"

Tests that utilize gestures:
* MultiFinger and TableView Tests benefit from waitFor... method calls.

-