Added option to select UIPickers based on accessibility identifier.
Previously the framework assumed there would only be one Picker on the screen at a time. When one wanted to change the value of the picker, KIF would search for all the pickers on the screen and pick the last one which doesn't make sense if a certain screen has multiple pickers. The old functionality still exists in the framework, but this addition allows us to now select a specific UIPicker or UIDatePicker based on accessibility identifier, and then change its value in our tests.
This also extends the existing UIPickerView in Test Host to a multi-component UIPickerView. Previously the function 'selectPickerViewRowWithTitle:inComponent' was tested on a UIDatePicker, but the new changes to selecting a UIPickerView or UIDatePicker are more specific in which object they select and tests should stick to 'selectDatePickerValue' in changing date picker values.