Elements don't become first responder fast enough
When tapping an element, there is a check that ensures it becomes the
first responder within a half second. This doesn't work for elements
such as popover controls however, which only become first responder
after the popover view finishes being presented (usually animated). In
that case, there is a race condition as to whether the element will
finish displaying before the polling interval completes.

The fix here is to ensure that we wait for any animations to complete
before trying to check that the element that has been tapped became
first responder.

One exception to this is when interacting with an accessibility element
and not a view. In this case, the backing view for the may still be
visible on screen, even though the element itself is not. An example of
this is web views, where this would cause the test `-[WebViewTests
testTappingLinks]` to fail.