Widen `firstResponder` searches to all windows
Previously searches for `UIWindow.firstResponder` were limited to the
`UIApplication.keyWindow`. In practice, it is very possible for users to
add additional valid windows that are not the key window and have first
responders. To accommodate those cases, this change broadens the search
for a first responder of an application to an array of first responders
found in `UIApplication.windowsWithKeyWindow`.