Address flaky tests in new CI configuration
* AccessibilityIdentifierTests tests

Fix `testEnteringTextIntoViewWithAccessibilityIdentifier` flaking

We're trying to tap on `Select All` while it is still being animated into view. The problem is two-fold:

1. The `longPress` action doesn't wait for animations to complete, it just uses a hardcoded .5s before
starting the next action.
2. Even after switching to use `waitForAnimationsToFinish`, the timeout defaults to .5s as the max to
wait for animations to complete.

To fix this, I switched `longPress` action to use `waitForAnimationsToFinish` and I upped the
default timeout to wait up to 2s for animations to complete. This seems to work much more reliably now.

* SystemTests.testMockingOpenURL

NSURL no longer parses '423543523454://', use a different protocol that starts with characters

* ExistTests

Wait a bit longer for the view controller presentation to complete

* TypingTests

In TapViewController, avoid race condition in resigning first responder and becoming first responder.
CI was flaking with a failure where "Deleted something., world" was found in the greeting text field.

* PullToRefreshTests

Leverage 'usingCurrentFrame' so we're not scrolling the table view around
while the UIRefreshControl is spinning, which immediately halts the refresh
spinner. Also, make sure that we're scrolled to the top of the test suite VC,
as otherwise the pull to refresh action might fail.