Introducing tryFindingViewInFrame() to avoid scroll (#1296)
Introduce `-[KIFUITestActor tryFindingViewInFrame]` and `-[KIFUIViewTestActor usingCurrentFrame]`.

This can be used to avoid scrolling when looking for a view. This can be helpful where implicit table/collection view scrolling that happens when attempting to find UI elements triggers unwanted behavior (pagination fetch, abort pull to refresh, etc).

This includes a behavior change in `isTappable` where it no longer automatically treats an element as tappable just for having a tap gesture recognizer on it, even though it might be occluded. See this discussion for more context:
https://github.com/kif-framework/KIF/pull/1296#discussion_r1696013400

---------

Authored-by: Simone Scionti <simone.scionti@qonto.com>