Update CI config to run on Xcode 14.3.1 and 15.4
Configuring CI to run in 4 configurations:
 - xcode 14.3.1, iPad Air (5th generation), iOS 16.4
 - xcode 14.3.1, iPhone 14, iOS 16.4
 - xcode 15.4, iPad Air (5th generation), iOS 17.5
 - xcode 15.4, iPhone 15, iOS 17.5

As part of this change, the CI script was updated to only run the
extra validation of the examples folder one one of the 4 shards.

This removes Carthage from the README as being officially supported.
I'd guess that is generally not being used. The CI script hasn't worked
since Xcode 12.