Up the default animation waiting timeout from .5s to 2s
It doesn't really make sense that the default animation wait time
is the same as the stabilization time. This effectively means the
behavior of scanning the UI hierarchy for animations in progress is
never utilized for most tests out of the box.

For what its worth, we've used a 2s animation wait time and .1s
stabilization time as our defaults for a very long time.

This should generally not slow down any tests if there weren't any
animations in progress. If there were animations in progress, this
is probably a worthwhile tradeoff to make things more reliable. This
can still be configured globally to be overidden.

As part of this change, use `waitForAnimationsToFinish` in place of
a static .5s sleep for the `longPress` action.

There will be a followup PR #1301 that should speed up some scenarios
where we're waiting longer than desired for animations to complete.