Ensure that the UITextEffectsWindow also gets layer speed multipliers applied
The UITextEffectsWindow directly uses a private initializer, bypassing the public initializer.
The `[UIWindow init]` implementation will call through to this private designated initializer,
so overriding both is unnecessary. I've added an assertion to the swizzle helpers to prevent
this from silently failing in the future if the private APIs change.