Improve detection of in progress animations
There were cases where we were not identifying animations that were still
in progress. For example, if I turn on the simulator slow animations setting,
we don't wait a sufficient amount of time for animations to complete. This
was consistently showing up when running -[TypingTests testEnteringTextIntoFirstResponder]`.

As this test was failing on Xcode 16 in CI, it seems like this change is
needed to get KIF working reliably.